some work on tests. tried to set debug on db2

+++ uninstall.sh
@@ -9,7 +9,7 @@
 ################################################################################
 
 ######################################################################~ START ##
-clear
+sudo clear
 echo "Uninstalling XML Disclosure"
 
 ################################################################~ READ CONFIG ##
@@ -19,24 +19,21 @@ echo "  . Reading configuration file"
 #####################################################################~ REMOVE ##
 echo "  . Removing unnecessary folders"
 rm -r class/* > $XD_UNST_LOG 2>&1
-rm -r jar/* > $XD_UNST_LOG 2>&1
-rm -r log/* > $XD_UNST_LOG 2>&1
+rm -r jar/* >> $XD_UNST_LOG 2>&1
+rm -r log/* >> $XD_UNST_LOG 2>&1
+rm -r $XD_DB2_FN/XMLDisclosure >> $XD_UNST_LOG 2>&1
 
-###################################################################~ DB2START ##
-echo "  . Ensuring DB2 is started"
-db2start \
-  >> $XD_UNST_LOG
-
-db2 \
-  connect reset \
-  > /dev/null
-
-#######################################################################~ DROP ##
-echo "  . Dropping the $XD_DB2_DB database"
-db2 \
-  drop database $XD_DB2_DB \
-  > $XD_UNST_LOG 2>&1
+########################################################################~ DB2 ##
+if [ "$1" = "all" ]; then
+  echo "  . Restarting DB2"
+  db2stop force >> $XD_UNST_LOG
+  db2start >> $XD_UNST_LOG
   
+  echo "  . Dropping the $XD_DB2_DB database"
+  db2 \
+    drop database $XD_DB2_DB \
+    >> $XD_UNST_LOG 2>&1
+fi
 ########################################################################~ EOF ##
 echo "XMLDisclosure processing COMPLETE"
 echo "Please refer to $XD_UNST_LOG for a detailed log file"
